requirements.txt
thermo
Pillow